% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehhR.R
\name{ehhR}
\alias{ehhR}
\title{A function to calculate EHH (Extended Haplotype Homozygosity) given a focal SNPs and a haplotype matrix.}
\usage{
ehhR(hapmatrix, site)
}
\arguments{
\item{hapmatrix}{A haplotype matrix with ones (ancestral) and twos (derived).}

\item{site}{A specific site.}
}
\value{
EHH (Extended Haplotype Homozygosity).
}
\description{
A function to calculate EHH (Extended Haplotype Homozygosity) given a focal SNPs and a haplotype matrix.
}
\examples{
exmtx=matrix(sample(2,80,replace=T),ncol=10,nrow=8)
ehhR(exmtx,2)
ehhR(exmtx,4)

}
\author{
Fei Xiang (\email{xf3087@gmail.com})
}

