% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehhsR.R
\name{ehhsR}
\alias{ehhsR}
\title{A function to calculate  EHHs (site Extended Haplotype Homozygosity) given a focal site and a haplotype matrix.}
\usage{
ehhsR(hapmatrix, site)
}
\arguments{
\item{hapmatrix}{A haplotype matrix with ones (ancestral) and twos (derived).}

\item{site}{A specific site.}
}
\value{
EHH (Extended Haplotype Homozygosity).
}
\description{
A function to calculate  EHHs (site Extended Haplotype Homozygosity) given a focal site and a haplotype matrix.
}
\examples{
exmtx=matrix(sample(2,80,replace=T),ncol=10,nrow=8)
ehhsR(exmtx,2)
ehhsR(exmtx,4)

}
\author{
Fei Xiang (\email{xf3087@gmail.com})
}

