% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flogliksingle.R
\name{flogliksingle}
\alias{flogliksingle}
\title{A function to calculate loglikelihood of a single locus model.}
\usage{
flogliksingle(parr, freqMatrix, rate, pos)
}
\arguments{
\item{parr}{Parameter vector.}

\item{freqmatrix}{Longitudinal frequency matrix of 4 nucleotides (A, C, G, T).}

\item{Pos}{The position of A, C, G, T under selection.}
}
\value{
loglikelihood of a single locus model.
}
\description{
A function to calculate loglikelihood of a single locus model.
}
\examples{
freq=cbind(D2=c(0,0,57,0),D3=c(0,0,44,0),D4=c(3,0,45,0))
rownames(freq)=c("a","c","g","t")
flogliksingle(c(0,1,1,1,1,1),freq,1/3*10^(-5),1)
library(minqa)
bobyqa(c(.1,.1,.1,.1,.1,.1),flogliksingle,upper=Inf,lower=0,freqMatrix=freq,rate=1/3*10^(-5),pos=1)

}
\author{
Fei Xiang (\email{xf3087@gmail.com})
}
