% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldR.R
\name{ldR}
\alias{ldR}
\title{A function to calculate Kelly's Zns statistic and the omega_max statistic given a haplotype matrix.}
\usage{
ldR(hapmatrix)
}
\arguments{
\item{hapmatrix}{A haplotype matrix with ones (ancestral) and twos (derived).}
}
\value{
Kelly's Zns statistic, the omega_max statistic and the site produces omega_max.
}
\description{
A function to calculate Kelly's Zns statistic and the omega_max statistic given a haplotype matrix.
}
\examples{
exmtx=matrix(sample(2,80,replace=T),ncol=10,nrow=8)
ldR(exmtx)

}
\author{
Fei Xiang (\email{xf3087@gmail.com})
}
