% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucdivR.R
\name{nucdivR}
\alias{nucdivR}
\title{A function to calculate the average hamming distance within each two populations, between two populations and Fst statistic.}
\usage{
nucdivR(hapmatrix1, hapmatrix2)
}
\arguments{
\item{hapmatrix1}{A haplotype matrix with ones (ancestral) and twos (derived).}

\item{hapmatrix2}{The other haplotype matrix with ones (ancestral) and twos (derived).}
}
\value{
The average hamming distance within each two populations, between two populations and Fst statistic.
}
\description{
A function to calculate the average hamming distance within each two populations, between two populations and Fst statistic.
}
\examples{
exmtx=matrix(sample(2,80,replace=T),ncol=10,nrow=8)
exmtx1=matrix(sample(2,80,replace=T),ncol=10,nrow=8)
nucdivR(exmtx,3-exmtx)
nucdivR(exmtx,exmtx1)

}
\author{
Fei Xiang (\email{xf3087@gmail.com})
}
