% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pileupFreq.R
\name{pileupFreq}
\alias{pileupFreq}
\title{Converts a pileup to a dataframe of nucleotide frequencies}
\usage{
pileupFreq(p)
}
\arguments{
\item{p}{a pileup}
}
\value{
Nucleotide frequencies (as a data frame)
}
\description{
\code{pileupFreq} converts the output of \code{pileup} to a dataframe of nucleotide frequencies.
}
\author{
Simon Frost (\email{sdwfrost@gmail.com})
}
