% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_bam_eff.R
\name{screen_bam_eff}
\alias{screen_bam_eff}
\title{An efficient function to identify sites of interestes from a list of longitudinal sequence file.}
\usage{
screen_bam_eff(file_list, ref = NULL, region, priorPA = 0.001)
}
\arguments{
\item{file_list}{A list of bam files.}

\item{ref}{A referece sequence.}

\item{region}{An optional region from the BAM file (as a \code{GRanges} object)}

\item{priorPA}{Prior assocation to possible models}
}
\value{
PPA association for each site.
}
\description{
An efficient function to identify sites of interestes from a list of longitudinal sequence file.
}
\examples{



p4305.d3=system.file("extdata/4305-d3.bam",package = "popseq")
p4305.d4=system.file("extdata/4305-d4.bam",package = "popseq")
p4305.d5=system.file("extdata/4305-d5.bam",package = "popseq")
ref.seg1=system.file("extdata/seg1.fas",package = "popseq")
ref.seg2=system.file("extdata/seg2.fas",package = "popseq")


pigbam=list(p4305.d3,p4305.d4,p4305.d5)


res1=screen_bam_eff(file_list=pigbam,ref=ref.seg1,region= GRanges("GQ166656_1",IRanges(1,2280)),priorPA=.001)

res2=screen_bam_eff(file_list=pigbam,ref=ref.seg2,priorPA=.001,region = GRanges("GQ166655_2",IRanges(1,2274)))







}
\author{
Fei Xiang (\email{xf3087@gmail.com})
}
