% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_bam_multi.R
\name{screen_bam_multi}
\alias{screen_bam_multi}
\title{An efficient function to identify sites of interestes from a list of longitudinal sequence files with parallel computing######}
\usage{
screen_bam_multi(file_list, ref = NULL, region, priorPA = 0.001, core = 1)
}
\arguments{
\item{file_list}{A list of bam files.}

\item{ref}{A referece sequence.}

\item{region}{An optional region from the BAM file (as a \code{GRanges} object).}

\item{priorPA}{Prior assocation to possible models.}

\item{core}{Number of cores for parallel computation.}
}
\value{
PPA association for each site.
}
\description{
An efficient function to identify sites of interestes from a list of longitudinal sequence files with parallel computing######
}
\examples{



p4305.d3=system.file("extdata/4305-d3.bam",package = "popseq")
p4305.d4=system.file("extdata/4305-d4.bam",package = "popseq")
p4305.d5=system.file("extdata/4305-d5.bam",package = "popseq")
ref=system.file("extdata/A_England_195_2009.fa",package = "popseq")



pigbam=list(p4305.d3,p4305.d4,p4305.d5)


res1=screen_bam_multi(file_list=pigbam,ref=ref,region= GRanges("GQ166656_1",IRanges(1,2280)),priorPA=.001,core=8)

res2=screen_bam_multi(file_list=pigbam,ref=ref,region = GRanges("GQ166655_2",IRanges(1,2274)),priorPA=.001,core=8)







}
\author{
Fei Xiang (\email{xf3087@gmail.com})
}
