% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_select.R
\name{screen_fasta}
\alias{screen_fasta}
\title{A function to identify sites of interestes from a list of longitudinal sequence file.}
\usage{
screen_fasta(file_list, ref = NULL, priorPA = 0.001)
}
\arguments{
\item{file_list}{A list of fasta files.}

\item{ref}{A referece sequence.}

\item{priorPA}{Prior assocation to possible models}
}
\value{
PPA association for each site.
}
\description{
A function to identify sites of interestes from a list of longitudinal sequence file.
}
\examples{
pig2.1=system.file("extdata/pig2.1",package = "popseq")
pig2.2=system.file("extdata/pig2.2",package = "popseq")
pig2.3=system.file("extdata/pig2.3",package = "popseq")

ref=system.file("extdata/regions.fas",package = "popseq")

pig2=list(pig2.1,pig2.2,pig2.3)

screen_select(pig2,ref=ref)

}
\author{
Fei Xiang (\email{xf3087@gmail.com})
}
